+++ js/data.js
@@ -1,0 +1,34 @@
+const OBJECT_COUNT = 25;
+
+const NAMES = [
+  'Кузьма',
+  'Прасковья',
+  'Инга',
+  'Шерлок',
+  'Феликс',
+  'Тамара',
+  'Яна',
+  'Святослав',
+  'Абрам',
+  'Глафира'
+];
+
+const DESCRIPTIONS = [
+  'важный',
+  'знаменитый',
+  'ласковый',
+  'вечный',
+  'смешанный',
+  'исполнительный',
+  'интенсивный',
+  'смеющийся',
+  'блестящий',
+  'утренний'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.'
+];
+
+export {OBJECT_COUNT, NAMES, DESCRIPTIONS, MESSAGES};
